Create a new intervention model to specify more complex and time variable exposure patterns (interventions).
This is done by adding a layer in between treatment and observation classes, the Timeseries model.

A treatment now receives observations and interventions to define a treatment. Both types (o/i) are defined as timeseries objects can thus be flexible over time. If the new database structure works out remains to be seen, when data are passed to the model.